Update init.sh and tht/overcloud-ceph-ansible.yaml to specify flavors
To ensure that the POC will always run with the correct node assignment,
I added the following to tht/overcloud-ceph-ansible.yaml:

  OvercloudControlFlavor: control
  OvercloudComputeFlavor: compute

This assumes that the following has been run before deployment:

 ironic node-update ceph-0 replace properties/capabilities=profile:compute,boot_option:local
 ironic node-update control-0 replace properties/capabilities=profile:control,boot_option:local

The above is now in init.sh so it will happen on a new POC install.

The profile is a capability assigned to both ironic node and nova flavor
to create a link between them. See `openstack flavor show compute` or the
following doc:

 https://docs.openstack.org/developer/tripleo-docs/advanced_deployment/profile_matching.html