Adding ceph-ansible-logs.sh
ceph-ansible-logs.sh will always return the logs from the last
ceph-ansible run. It uses a heurestic of the first line of any
ceph-ansible-3.0.x run starting with the same deprecation warning.
It then does simple aritmetic based on line that occurred and the
number of lines in the log and tails only those lines.

This is just a debug tool.